move hardware-related headers to their own folder
