+++ src/include/hardware/gpu_info.h
@@ -3,7 +3,7 @@
 
 #define BENJI_USE_SYS_INFO_UTILS
 
-#include "utils.h"
+#include "../utils.h"
 
 typedef struct _BENJI_GPU_INFO {
     char* name;
